// src/sms/enums/sms.enums.ts
export enum SmsReceiver {
    ALL = 'all',
    INDIVIDUAL_USERS = 'individualUsers',
    SPECIFIC_COURSES = 'specificCourses',
    BATCH_COURSE_PARENTS = 'batchCourseParents',
    COURSE_CATEGORIES = 'courseCategories',
    HAVING_DUES_OF_SPECIFIC_COURSES = 'havingDuesOfSpecifCourses',
    PHONE_NUMBERS = 'phoneNumbers',
    FULL_PAID_COURSES = 'fullPaidCourses',
}